+++ src/commands/customToCustomCommand.ts
@@ -1,0 +1,108 @@
+/*!
+ * ---------------------------------------------------------------------------------------------
+ *  Copyright (c) Leo Hanisch. All rights reserved.
+ *  Licensed under the MIT License. See LICENSE.txt in the project root for license information.
+ * --------------------------------------------------------------------------------------------
+ */
+
+'use strict';
+
+import { InputBoxStep } from '../step/inputBoxStep';
+import { MultiStepHandler } from '../step/multiStepHandler';
+import { QuickPickStep } from '../step/quickPickStep';
+import { StepResult } from '../step/stepResult';
+import { InputFlowAction } from '../util/InputFlowAction';
+import { CustomCommandBase } from './customCommandBase';
+
+class CustomToCustomCommand extends CustomCommandBase {
+
+    private readonly title: string = 'Custom → Custom';
+
+    public async execute() {
+        const preSelection = this.isInputSelected();
+        let loopResult: StepResult = new StepResult(InputFlowAction.Continue, preSelection);
+        do {
+            let rawInput = loopResult.value;
+            let customSourceFormat: string;
+            let customTargetFormat: string;
+
+            if (!this._stepHandler) {
+                this.initialize();
+            }
+
+            if (loopResult.action === InputFlowAction.Back) {
+                [customSourceFormat, rawInput, customTargetFormat] =
+                    await this._stepHandler.run(this._ignoreFocusOut, rawInput, -1);
+            } else {
+                [customSourceFormat, rawInput, customTargetFormat] =
+                    await this._stepHandler.run(this._ignoreFocusOut, rawInput);
+            }
+
+            if (!rawInput || !customSourceFormat || !customTargetFormat) {
+                break;
+            }
+
+            const result = this._timeConverter.customToCustom(rawInput, customSourceFormat, customTargetFormat);
+
+            let inserted: boolean = false;
+            if (this._insertConvertedTime) {
+                inserted = await this.insert(result);
+            }
+            const titlePostfix = (inserted ? ': Inserted Result' : ': Result') + ' (' + customTargetFormat + ')';
+
+            loopResult = await this._resultBox.show(
+                'Input: ' + rawInput + ' (Format: ' + customSourceFormat + ')',
+                this.title + titlePostfix,
+                result,
+                this.insert,
+                this._ignoreFocusOut);
+        } while (loopResult.action === InputFlowAction.Back
+            || (!this._hideResultViewOnEnter && loopResult.action === InputFlowAction.Continue));
+    }
+
+    private initialize(): void {
+        const alternativeCustomFormatStep1 = new InputBoxStep(
+            'E.g.: YYYY/MM/DD',
+            'Insert custom format',
+            this.title,
+            'Ensure you enter a custom momentjs format.',
+            (input) => input ? true : false,
+            false,
+            true);
+        const alternativeCustomFormatStep2 = new InputBoxStep(
+            'E.g.: YYYY/MM/DD',
+            'Insert custom format',
+            this.title,
+            'Ensure you enter a custom momentjs format.',
+            (input) => input ? true : false,
+            false,
+            true);
+
+        const getCustomSourceFormatStep = new QuickPickStep(
+            'Select custom source format.',
+            this.title,
+            this._customTimeFormatOptions,
+            { label: 'Other Format...' },
+            alternativeCustomFormatStep1);
+        const getTimeOfCustomFormat = new InputBoxStep(
+            '$(prior-result)',
+            'Insert time of format: $(prior-result)',
+            this.title,
+            'Ensure the inserted time has format: $(prior-result)',
+            this._timeConverter.isValidCustom,
+            true);
+        const getCustomTargetFormatStep = new QuickPickStep(
+            'Select custom target format.',
+            this.title,
+            this._customTimeFormatOptions,
+            { label: 'Other Format...' },
+            alternativeCustomFormatStep2);
+
+        this._stepHandler = new MultiStepHandler();
+        this._stepHandler.registerStep(getCustomSourceFormatStep);
+        this._stepHandler.registerStep(getTimeOfCustomFormat);
+        this._stepHandler.registerStep(getCustomTargetFormatStep);
+    }
+}
+
+export { CustomToCustomCommand };