+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to the "timing" extension will be documented in this file. This project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+* **Added** `timing.customToCustom` command to convert a custom time to another custom format.
+
 ## 2019-01-28 - v1.3.0
 * **Added** _weeks_ to the readable duration conversion.
 * **Added** `timing.isoDurationToEpoch` command to convert an ISO 8601 duration to an epoch timestamp.